refactor: remove redundant lesson management pages and standardize import paths for certificate components
