+++ frontend/components/admin/quiz-manager-panel.tsx
@@ -8,6 +8,7 @@ import {
   PointerSensor,
   useSensor,
   useSensors,
+  type DragEndEvent,
 } from "@dnd-kit/core"
 import {
   arrayMove,
@@ -182,7 +183,7 @@ export function QuizManagerPanel({
     useSensor(KeyboardSensor, { coordinateGetter: sortableKeyboardCoordinates })
   )
 
-  const handleDragEnd = (event: { active: { id: string }; over: { id: string } | null }) => {
+  const handleDragEnd = (event: DragEndEvent) => {
     if (!event.over) return
     const oldIdx = questions.findIndex((q) => (q._sortId || q.id) === event.active.id)
     const newIdx = questions.findIndex((q) => (q._sortId || q.id) === event.over!.id)
