+++ frontend/contexts/user-context.tsx
@@ -189,11 +189,7 @@ export function UserProvider({ children }: { children: React.ReactNode }) {
     setUserStatsLoading(true);
     setUserStatsError(false);
     try {
-      const data = await api.get<{
-        courseCount: number;
-        certificateCount: number;
-        xp: number;
-      }>("/users/me/stats");
+      const data = await api.get<UserStats>("/users/me/stats");
       setUserStats(data);
     } catch {
       setUserStatsError(true);
